--- Zylo/UI/babel.config.json
@@ -1,14 +1,0 @@
-{
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": "10"
-          }
-        }
-      ],
-      "@babel/preset-react"
-    ],
-    "plugins": ["@babel/plugin-proposal-class-properties"]
-}
