--- Zylo/UI/src/components/App.js
@@ -1,14 +1,0 @@
-import React, { Component} from "react";
-import { render } from 'react-dom';
-
-function App() {
-    return (
-        <div className="App">
-            <header className="App-header">
-                <h1>Welcome to Zylo</h1>
-            </header>
-        </div>
-    );
-}
-  
-export default App;
