--- Zylo/UI/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import index
-
-urlpatterns = [
-    path('', index)
-]
