--- Zylo/zyloapp/models.py
@@ -1,19 +1,0 @@
-from django.db import models
-import string
-import random
-
-def generate_unique_code():
-    length = 8
-    while True:
-        code = ''.join(random.choices(string.ascii_uppercase, k=length))
-        if Room.objects.filter(code=code).count() == 0:
-            break
-    return code
-
-# Create your models here.
-class Room(models.Model):
-    code = models.CharField(max_length=10, default="", unique=True)
-    host = models.CharField(max_length=50, unique=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-
