--- Zylo/zyloapp/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from .models import Room
-
-class RoomSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Room
-        fields = ('id', 'code', 'host', 'created_at')
