+++ zylo-backend/backend/accounts/models.py
@@ -1,0 +1,6 @@
+from django.db import models
+from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
+
+# Create your models here.
+class UserManager:
+
