+++ zylo-backend/backend/requirements.txt
@@ -1,0 +1,2 @@
+django==5.1.1
+djangorestframework==3.15.2
