+++ zylo-frontend/ui/src/App.js
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      Zylo Chat App
+    </div>
+  );
+}
+
+export default App;
